pyproject.toml
[project]
name = "adk-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.9.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
]
